backend/requirements.txt
Flask
Flask-Cors
Pillow
numpy
google-generativeai
python-dotenv